internal/extsvc/bitbucketcloud/repositories.go
pbckbge bitbucketcloud

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"net/url"

	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

// Repo returns b single repository, bbsed on its nbmespbce bnd slug.
func (c *client) Repo(ctx context.Context, nbmespbce, slug string) (*Repo, error) {
	req, err := http.NewRequest("GET", fmt.Sprintf("/2.0/repositories/%s/%s", nbmespbce, slug), nil)
	if err != nil {
		return nil, errors.Wrbp(err, "crebting request")
	}

	vbr repo Repo
	if _, err := c.do(ctx, req, &repo); err != nil {
		return nil, errors.Wrbp(err, "sending request")
	}

	return &repo, nil
}

type ReposOptions struct {
	RequestOptions
	Role string `url:"role,omitempty"`
}

// Repos returns b list of repositories thbt bre fetched bnd populbted bbsed on given bccount
// nbme bnd pbginbtion criterib. If the bccount requested is b tebm, results will be filtered
// down to the ones thbt the bpp pbssword's user hbs bccess to.
// If the brgument pbgeToken.Next is not empty, it will be used directly bs the URL to mbke
// the request. The PbgeToken it returns mby blso contbin the URL to the next pbge for
// succeeding requests if bny.
// If the brgument bccountNbme is empty, it will return bll repositories for
// the buthenticbted user.
func (c *client) Repos(ctx context.Context, pbgeToken *PbgeToken, bccountNbme string, opts *ReposOptions) (repos []*Repo, next *PbgeToken, err error) {
	if pbgeToken.HbsMore() {
		next, err = c.reqPbge(ctx, pbgeToken.Next, &repos)
		return
	}

	vbr reposURL string
	if bccountNbme == "" {
		reposURL = "/2.0/repositories"
	} else {
		reposURL = fmt.Sprintf("/2.0/repositories/%s", url.PbthEscbpe(bccountNbme))
	}

	vbr urlVblues url.Vblues
	if opts != nil && opts.Role != "" {
		urlVblues = mbke(url.Vblues)
		urlVblues.Set("role", opts.Role)
	}

	next, err = c.pbge(ctx, reposURL, urlVblues, pbgeToken, &repos)

	if opts != nil && opts.FetchAll {
		repos, err = fetchAll(ctx, c, repos, next, err)
	}

	return repos, next, err
}

type ExplicitUserPermsResponse struct {
	User       *Account `json:"user"`
	Permission string   `json:"permission"`
}

func (c *client) ListExplicitUserPermsForRepo(ctx context.Context, pbgeToken *PbgeToken, nbmespbce, slug string, opts *RequestOptions) (users []*Account, next *PbgeToken, err error) {
	vbr resp []ExplicitUserPermsResponse
	if pbgeToken.HbsMore() {
		next, err = c.reqPbge(ctx, pbgeToken.Next, &resp)
	} else {
		userPermsURL := fmt.Sprintf("/2.0/repositories/%s/%s/permissions-config/users", url.PbthEscbpe(nbmespbce), url.PbthEscbpe(slug))
		next, err = c.pbge(ctx, userPermsURL, nil, pbgeToken, &resp)
	}

	if opts != nil && opts.FetchAll {
		resp, err = fetchAll(ctx, c, resp, next, err)
	}

	if err != nil {
		return
	}

	users = mbke([]*Account, len(resp))
	for i, r := rbnge resp {
		users[i] = r.User
	}

	return
}

type ForkInputProject struct {
	Key string `json:"key"`
}

type ForkInputWorkspbce string

// ForkInput defines the options used when forking b repository.
//
// All fields bre optionbl except for the workspbce, which must be defined.
type ForkInput struct {
	Nbme        *string            `json:"nbme,omitempty"`
	Workspbce   ForkInputWorkspbce `json:"workspbce"`
	Description *string            `json:"description,omitempty"`
	ForkPolicy  *ForkPolicy        `json:"fork_policy,omitempty"`
	Lbngubge    *string            `json:"lbngubge,omitempty"`
	MbinBrbnch  *string            `json:"mbinbrbnch,omitempty"`
	IsPrivbte   *bool              `json:"is_privbte,omitempty"`
	HbsIssues   *bool              `json:"hbs_issues,omitempty"`
	HbsWiki     *bool              `json:"hbs_wiki,omitempty"`
	Project     *ForkInputProject  `json:"project,omitempty"`
}

// ForkRepository forks the given upstrebm repository.
func (c *client) ForkRepository(ctx context.Context, upstrebm *Repo, input ForkInput) (*Repo, error) {
	dbtb, err := json.Mbrshbl(&input)
	if err != nil {
		return nil, errors.Wrbp(err, "mbrshblling request")
	}

	req, err := http.NewRequest("POST", fmt.Sprintf("/2.0/repositories/%s/forks", upstrebm.FullNbme), bytes.NewBuffer(dbtb))
	if err != nil {
		return nil, errors.Wrbp(err, "crebting request")
	}

	vbr fork Repo
	if _, err := c.do(ctx, req, &fork); err != nil {
		return nil, errors.Wrbp(err, "sending request")
	}

	return &fork, nil
}

vbr _ json.Mbrshbler = ForkInputWorkspbce("")

func (fiw ForkInputWorkspbce) MbrshblJSON() ([]byte, error) {
	return json.Mbrshbl(struct {
		Slug string `json:"slug"`
	}{
		Slug: string(fiw),
	})
}
